refactor: move core functionality to lynx module
