+++ src/main.rs
@@ -1,3 +1,4 @@
+mod lynx;
 mod wal;
 
 use axum::{
@@ -9,19 +10,9 @@ use axum::{
 };
 use clap::Parser;
 use serde::{Deserialize, Serialize};
-use std::{
-    collections::{BTreeMap, btree_map::Entry},
-    sync::{Arc, Mutex},
-};
-use std::{net::SocketAddr, path::PathBuf};
-use wal::TagValue;
+use std::{net::SocketAddr, path::PathBuf, sync::Arc};
 
-use crate::wal::{Wal, WriteRequest};
-
-/// Time format string for daily partition keys.
-///
-/// This means that data is partitioned by day at all times currently.
-const DAILY_PARTITION: &str = "%Y-%m-%d";
+use crate::{lynx::Lynx, wal::WriteRequest};
 
 #[derive(Parser, Debug)]
 struct Args {
@@ -36,20 +27,8 @@ struct Args {
 }
 
 struct AppState {
-    /// Write-ahead log to provide durable writes for incoming data.
-    ///
-    /// Data MUST be appended to the WAL before making its way into the
-    /// in-memory buffer.
-    wal: Mutex<Wal>,
-    /// In-memory structure which makes the durable writes queryable.
-    buffer: Arc<Mutex<BTreeMap<String, BTreeMap<String, Measurements>>>>,
-}
-
-#[derive(Default, Debug)]
-struct Measurements {
-    timestamps: Vec<u64>,
-    tags: Vec<(String, TagValue)>,
-    values: Vec<String>,
+    /// An in-memory, durable, time-series database.
+    lynx: Lynx,
 }
 
 #[derive(Deserialize)]
@@ -72,53 +51,13 @@ async fn write_handler(
     State(state): State<Arc<AppState>>,
     Json(payload): Json<WriteRequest>,
 ) -> impl IntoResponse {
-    match state.wal.lock().unwrap().write(payload.clone()) {
-        Ok(_) => {}
+    match state.lynx.write(payload) {
+        Ok(_) => StatusCode::OK,
         Err(e) => {
             eprintln!("{e:?}");
-            return StatusCode::INTERNAL_SERVER_ERROR;
+            StatusCode::INTERNAL_SERVER_ERROR
         }
     }
-    let mut buffer_guard = state.buffer.lock().unwrap();
-    match buffer_guard.entry(payload.namespace) {
-        Entry::Vacant(vacant) => {
-            let inbound_time = chrono::DateTime::from_timestamp_micros(payload.timestamp as i64)
-                .expect("timestamps are currently assumed to be microseconds");
-
-            let partition_key = inbound_time.format(DAILY_PARTITION);
-            vacant.insert(BTreeMap::from_iter([(
-                partition_key.to_string(),
-                Measurements {
-                    timestamps: vec![payload.timestamp],
-                    tags: payload.tags,
-                    values: vec![payload.value],
-                },
-            )]));
-        }
-        Entry::Occupied(mut buffer_entry) => {
-            let partitions = buffer_entry.get_mut();
-
-            let inbound_time = chrono::DateTime::from_timestamp_micros(payload.timestamp as i64)
-                .expect("timestamps are currently assumed to be microseconds");
-            let partition_key = inbound_time.format(DAILY_PARTITION);
-            match partitions.entry(partition_key.to_string()) {
-                Entry::Vacant(init) => {
-                    init.insert(Measurements {
-                        timestamps: vec![payload.timestamp],
-                        tags: payload.tags,
-                        values: vec![payload.value],
-                    });
-                }
-                Entry::Occupied(mut buffered_measurements) => {
-                    let buffered_measurements = buffered_measurements.get_mut();
-                    buffered_measurements.timestamps.push(payload.timestamp);
-                    buffered_measurements.tags.extend(payload.tags);
-                    buffered_measurements.values.push(payload.value);
-                }
-            };
-        }
-    };
-    StatusCode::OK
 }
 
 async fn query_handler(
@@ -139,8 +78,7 @@ async fn main() {
     let args = Args::parse();
 
     let state = Arc::new(AppState {
-        wal: Mutex::new(Wal::new(&args.wal_directory, args.wal_max_segment_size)),
-        buffer: Arc::new(Mutex::new(BTreeMap::new())),
+        lynx: Lynx::new(&args.wal_directory, args.wal_max_segment_size),
     });
 
     let app = Router::new()
